Double processing

+++ Task_1_1_3/src/main/java/ru/nsu/lyskov/Add.java
@@ -25,7 +25,7 @@ public class Add extends Expression {
     }
 
     @Override
-    public int eval(Map<String, Integer> variables) {
+    public double eval(Map<String, Double> variables) {
         return left.eval(variables) + right.eval(variables);
     }
 }
