+++ Task_1_1_3/src/main/java/ru/nsu/lyskov/Mul.java
@@ -26,7 +26,7 @@ public class Mul extends Expression {
     }
 
     @Override
-    public int eval(Map<String, Integer> variables) {
+    public double eval(Map<String, Double> variables) {
         return left.eval(variables) * right.eval(variables);
     }
 }
